+++ js/message.js
@@ -1,3 +1,5 @@
+import { isEscEvent } from './util.js';
+
 // Показываем сообщение
 function showMessage(selector) {
   const template = document.querySelector(selector).content.querySelector('div');
@@ -13,4 +15,27 @@ function removeMessage(elementForRemove) {
   elementForRemove.remove();
 }
 
-export { showMessage, removeMessage };
+// Сообщение об ошибки при получении данных
+function messageForErrorGetData() {
+  const message = showMessage('#error-fetch');
+  const buttonRepeat = message.querySelector('.error__button');
+  message.addEventListener('click', removeModal);
+  buttonRepeat.addEventListener('click', removeModal);
+  document.addEventListener('keydown', onEscKeydown);
+}
+
+// Удаление сообщения
+function removeModal() {
+  removeMessage(document.querySelector('main > [data-modal="message"]'));
+  document.removeEventListener('keydown', onEscKeydown);
+}
+
+// Проверка на нажатие Esc
+function onEscKeydown(evt) {
+  if (isEscEvent(evt)) {
+    evt.preventDefault();
+    removeModal();
+  }
+}
+
+export { showMessage, removeMessage, messageForErrorGetData };
